Loco Foco main activity xml updated

+++ app/src/main/java/com/example/locofoco/TimerService.kt
@@ -11,7 +11,7 @@ class TimerService: Service() {
     private val timer = Timer()
 
     override fun onStartCommand(intent: Intent, flags: Int, startId: Int): Int {
-        val time = intent.getDoubleExtra(TIME_EXTRA, 0.0)
+        val time = intent.getIntExtra(TIME_EXTRA, 0)
         timer.scheduleAtFixedRate(TimeTask(time), 0, 1000)
         return START_NOT_STICKY
 
@@ -22,7 +22,7 @@ class TimerService: Service() {
         super.onDestroy()
     }
 
-    private inner class TimeTask(private var time: Double) : TimerTask() {
+    private inner class TimeTask(private var time: Int) : TimerTask() {
         override fun run() {
             val intent = Intent(TIMER_UPDATED)
             time --
